+++ packages/ui/src/entries/empty.ts
@@ -1,0 +1,1 @@
+export { default as FEmpty } from '@/components/extensions/FEmpty.vue'